refactor: don't use protocol classes in client API
